requirements.txt
uvicorn
fastapi
pydentic
pymongo
python-multipart
pyjwt
werkzeug